# herecrm-pwa-openapi.yaml
openapi: 3.1.0
info:
  title: HereCRM PWA API
  version: 1.0.0
paths:
...
  /api/v1/pwa/onboarding/choice/:
    post:
      tags:
      - pwa-onboarding
      summary: Process Onboarding Choice
      description: "Handle the user's choice to either keep their auto-created business\
        \ \nor join an existing one via invite code."
      operationId: process_onboarding_choice_api_v1_pwa_onboarding_choice__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnboardingChoice'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/pwa/onboarding/choice:
    post:
      tags:
      - pwa-onboarding
      summary: Process Onboarding Choice
      description: "Handle the user's choice to either keep their auto-created business\
        \ \nor join an existing one via invite code."
      operationId: process_onboarding_choice_api_v1_pwa_onboarding_choice_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnboardingChoice'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/pwa/requests/:
    get:
      tags:
      - pwa-requests
      summary: List Requests
      operationId: list_requests_api_v1_pwa_requests__get
      parameters:
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: urgency
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Urgency
      - name: min_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Min Date
      - name: max_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Max Date
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RequestSchema'
                title: Response List Requests Api V1 Pwa Requests  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - pwa-requests
      summary: Create Request
      operationId: create_request_api_v1_pwa_requests__post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/pwa/requests/{request_id}:
    get:
      tags:
      - pwa-requests
      summary: Get Request
      operationId: get_request_api_v1_pwa_requests__request_id__get
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: integer
          title: Request Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - pwa-requests
      summary: Update Request
      operationId: update_request_api_v1_pwa_requests__request_id__patch
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: integer
          title: Request Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - pwa-requests
      summary: Delete Request
      operationId: delete_request_api_v1_pwa_requests__request_id__delete
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: integer
          title: Request Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/pwa/marketing/:
    get:
      tags:
      - pwa-marketing
      summary: List Campaigns
      operationId: list_campaigns_api_v1_pwa_marketing__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/CampaignSchema'
                type: array
                title: Response List Campaigns Api V1 Pwa Marketing  Get
    post:
      tags:
      - pwa-marketing
      summary: Create Campaign
      operationId: create_campaign_api_v1_pwa_marketing__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampaignCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/pwa/marketing/preview:
    post:
      tags:
      - pwa-marketing
      summary: Preview Audience
      operationId: preview_audience_api_v1_pwa_marketing_preview_post
      parameters:
      - name: query
        in: query
        required: false
        schema:
          type: string
          default: all
          title: Query
      - name: channel
        in: query
        required: false
        schema:
          type: string
          default: WHATSAPP
          title: Channel
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/pwa/marketing/{campaign_id}:
    get:
      tags:
      - pwa-marketing
      summary: Get Campaign
      operationId: get_campaign_api_v1_pwa_marketing__campaign_id__get
      parameters:
      - name: campaign_id
        in: path
        required: true
        schema:
          type: integer
          title: Campaign Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/pwa/marketing/{campaign_id}/execute:
    post:
      tags:
      - pwa-marketing
      summary: Execute Campaign
      operationId: execute_campaign_api_v1_pwa_marketing__campaign_id__execute_post
      parameters:
      - name: campaign_id
        in: path
        required: true
        schema:
          type: integer
          title: Campaign Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CampaignChannel:
      type: string
      enum:
      - WHATSAPP
      - email
      - SMS
      title: CampaignChannel
    CampaignCreate:
      properties:
        name:
          type: string
          maxLength: 100
          title: Name
        channel:
          $ref: '#/components/schemas/CampaignChannel'
        body:
          type: string
          maxLength: 5000
          title: Body
        subject:
          anyOf:
          - type: string
            maxLength: 200
          - type: 'null'
          title: Subject
        recipient_query:
          type: string
          maxLength: 500
          title: Recipient Query
          default: all
      type: object
      required:
      - name
      - channel
      - body
      title: CampaignCreate
    CampaignSchema:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          maxLength: 100
          title: Name
        channel:
          $ref: '#/components/schemas/CampaignChannel'
        status:
          $ref: '#/components/schemas/CampaignStatus'
        total_recipients:
          type: integer
          title: Total Recipients
        sent_count:
          type: integer
          title: Sent Count
        failed_count:
          type: integer
          title: Failed Count
        created_at:
          type: string
          format: date-time
          title: Created At
        sent_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Sent At
      type: object
      required:
      - id
      - name
      - channel
      - status
      - total_recipients
      - sent_count
      - failed_count
      - created_at
      title: CampaignSchema
    CampaignStatus:
      type: string
      enum:
      - DRAFT
      - SCHEDULED
      - SENDING
      - COMPLETED
      - FAILED
      - CANCELLED
      title: CampaignStatus
...
    CustomerSchema:
      properties:
        id:
          type: integer
          title: Id
        name:
          anyOf:
          - type: string
            maxLength: 200
          - type: 'null'
          title: Name
        first_name:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: First Name
        last_name:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Last Name
        phone:
          anyOf:
          - type: string
            pattern: (^\+?[1-9]\d{1,14}$|^$)
          - type: 'null'
          title: Phone
        email:
          anyOf:
          - type: string
            pattern: (^[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+$|^$)
          - type: 'null'
          title: Email
        street:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
          title: Street
        city:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: City
        postal_code:
          anyOf:
          - type: string
            maxLength: 20
          - type: 'null'
          title: Postal Code
        latitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Latitude
        longitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Longitude
        pipeline_stage:
          $ref: '#/components/schemas/PipelineStage'
        job_count:
          type: integer
          title: Job Count
          default: 0
        total_value:
          type: number
          title: Total Value
          default: 0.0
      type: object
      required:
      - id
      - pipeline_stage
      title: CustomerSchema
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LineItemCreate:
      properties:
        description:
          type: string
          maxLength: 500
          title: Description
        quantity:
          type: number
          title: Quantity
          default: 1.0
        unit_price:
          type: number
          title: Unit Price
        service_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Service Id
      type: object
      required:
      - description
      - unit_price
      title: LineItemCreate
    LineItemSchema:
      properties:
        id:
          type: integer
          title: Id
        description:
          type: string
          maxLength: 500
          title: Description
        quantity:
          type: number
          title: Quantity
        unit_price:
          type: number
          title: Unit Price
        total_price:
          type: number
          title: Total Price
        service_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Service Id
      type: object
      required:
      - id
      - description
      - quantity
      - unit_price
      - total_price
      title: LineItemSchema
...
    OnboardingChoice:
      properties:
        choice:
          type: string
          title: Choice
        invite_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Invite Code
        business_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Business Name
      type: object
      required:
      - choice
      title: OnboardingChoice
...
    PipelineStage:
      type: string
      enum:
      - NEW_LEAD
      - NOT_CONTACTED
      - CONTACTED
      - QUOTED
      - CONVERTED_ONCE
      - CONVERTED_RECURRENT
      - NOT_INTERESTED
      - LOST
      title: PipelineStage
...
    RequestCreate:
      properties:
        customer_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Customer Id
        description:
          type: string
          maxLength: 2000
          title: Description
        urgency:
          $ref: '#/components/schemas/Urgency'
          default: MEDIUM
        expected_value:
          anyOf:
          - type: number
          - type: 'null'
          title: Expected Value
        subtotal:
          anyOf:
          - type: number
          - type: 'null'
          title: Subtotal
          default: 0.0
        tax_amount:
          anyOf:
          - type: number
          - type: 'null'
          title: Tax Amount
          default: 0.0
        tax_rate:
          anyOf:
          - type: number
          - type: 'null'
          title: Tax Rate
          default: 0.0
        items:
          anyOf:
          - items:
              $ref: '#/components/schemas/LineItemCreate'
            type: array
          - type: 'null'
          title: Items
        follow_up_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Follow Up Date
        customer_details:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Customer Details
      type: object
      required:
      - description
      title: RequestCreate
    RequestSchema:
      properties:
        id:
          type: integer
          title: Id
        description:
          type: string
          maxLength: 2000
          title: Description
        status:
          $ref: '#/components/schemas/RequestStatus'
        urgency:
          $ref: '#/components/schemas/Urgency'
        expected_value:
          anyOf:
          - type: number
          - type: 'null'
          title: Expected Value
        subtotal:
          type: number
          title: Subtotal
          default: 0.0
        tax_amount:
          type: number
          title: Tax Amount
          default: 0.0
        tax_rate:
          type: number
          title: Tax Rate
          default: 0.0
        follow_up_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Follow Up Date
        customer_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Customer Id
        customer_details:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Customer Details
        created_at:
          type: string
          format: date-time
          title: Created At
        customer:
          anyOf:
          - $ref: '#/components/schemas/CustomerSchema'
          - type: 'null'
        line_items:
          items:
            $ref: '#/components/schemas/LineItemSchema'
          type: array
          title: Line Items
          default: []
      type: object
      required:
      - id
      - description
      - status
      - urgency
      - created_at
      title: RequestSchema
    RequestStatus:
      type: string
      enum:
      - PENDING
      - CONVERTED
      - DISMISSED
      - COMPLETED
      - LOGGED
      title: RequestStatus
    RequestUpdate:
      properties:
        description:
          anyOf:
          - type: string
            maxLength: 2000
          - type: 'null'
          title: Description
        status:
          anyOf:
          - $ref: '#/components/schemas/RequestStatus'
          - type: 'null'
        urgency:
          anyOf:
          - $ref: '#/components/schemas/Urgency'
          - type: 'null'
        expected_value:
          anyOf:
          - type: number
          - type: 'null'
          title: Expected Value
        subtotal:
          anyOf:
          - type: number
          - type: 'null'
          title: Subtotal
        tax_amount:
          anyOf:
          - type: number
          - type: 'null'
          title: Tax Amount
        tax_rate:
          anyOf:
          - type: number
          - type: 'null'
          title: Tax Rate
        items:
          anyOf:
          - items:
              $ref: '#/components/schemas/LineItemCreate'
            type: array
          - type: 'null'
          title: Items
        follow_up_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Follow Up Date
        customer_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Customer Id
      type: object
      title: RequestUpdate
...
    Urgency:
      type: string
      enum:
      - LOW
      - MEDIUM
      - HIGH
      title: Urgency
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...